Crate domain

A DNS library for Rust.

This crate provides a wide range of modules related to the Domain Name System. Currently, these are:

fundamental types, traits, and implementations for dealing with DNS data through the modules bits and iana,

parsing of master file data (aka zonefiles) in master,

data and master file access for various resource record types in rdata,

an asynchronous stub resolver implementation for querying the DNS in resolv.

Modules

bits
DNS data.

iana
IANA Definitions for DNS.

master
Reading and writing of master files.

rdata
Resource data implementations.

resolv
An asynchronous stub resolver.

utils
Various utility modules.