Struct domain::bits::charstr::CharStrBuf

pub struct CharStrBuf { /* fields omitted */ }

An owned, mutable DNS character string.

This type adds the methods push(), extend_from_slice(), and extend_from_str() to the methods from CharStr and [u8] to which it derefs (transitively). Note that the Extend trait is not implemented since adding data to a character string may fail if the string would become too long.

In addition through the usual suspects, values can be created from strings via the FromStr trait as well as by parsing from master file data through the scan() function.

Methods

impl CharStrBuf

Creation and Conversion

fn from_vec(vec: Vec<u8>) -> Option<Self>

Creates a new character string using the given vec.

Returns Some(_) if the vector was at most 255 long or None otherwise.

fn new() -> Self

Creates a new empty character string.

fn scan<S: Scanner>(scanner: &mut S) -> ScanResult<Self>

Scans a new character string from master data.

fn into_vec(self) -> Vec<u8>

Trades the character string for its raw bytes vector.

impl CharStrBuf

Manipulations.

fn push(&mut self, ch: u8) -> Result<(), PushError>

Appends an octet to the end of the character string.

The method will fail if there isn’t room for additional data.

fn extend_from_slice<B: AsRef<[u8]>>(&mut self,
bytes: B)
-> Result<(), PushError>

Extends the character string with the contents of a bytes slice.

The method will fail if there isn’t room for adding the complete contents of the slice.

fn extend_from_str(&mut self, s: &str) -> Result<(), FromStrError>

Extends the character string with the contents of a Rust string.

The string must contain the master data representation of a character string either as a regular word without white-space or a single quoted entity. If the string cannot be converted, the function fails with the appropriate error. If adding the resulting data would exceed the length limit of the character string, the method fails with FromStrError::LongString.