Struct domain::bits::compose::ComposeSnapshot

pub struct ComposeSnapshot { /* fields omitted */ }

A snapshot of a composer’s state.

This type is actually a composer all by itself, that is, it implements AsMut<Composer>. However, if necessary, it can be rolled back to the state the composer had when it was created forgetting about all changes made since.

This process currently is not transitive. While a new snapshot can be created from both a composer and a snapshot, when you roll back you get a composer. Ie., you can only roll back once. This is enough for the purpose, reuse of messages when transitioning through several servers in a resolver.

Methods

impl ComposeSnapshot

fn new(composer: Composer) -> Self

Creates a new snapshot from the given composer.

fn rewind(&mut self)

Rewinds the state to when the snapshot was taken.

fn rollback(self) -> Composer

Rolls back to a composer with state as when the snapshot was taken.

fn commit(self) -> Composer

Trades in the snapshot to a composer with all changes commited.