Module domain::bits::header

The header of a DNS message.

Each DNS message starts with a twelve octet long header section containing some general information related to the message as well as the number of records in each of its four sections. Its content and format are defined in section 4.1.1 of RFC 1035.

In order to reflect the fact that changing the section counts may invalidate the rest of the message whereas the other elements of the header section can safely be modified, the header section has been split into two separate types: Header contains the ‘safe’ part at the beginning of the section and HeaderCounts contains the section counts. In addition, the HeaderSection type wraps both of them into a single type.

Structs

Header
The first part of the header of a DNS message.

HeaderCounts
The section count part of the header section of a DNS message.

HeaderSection
The complete header section of a DNS message.