Struct domain::bits::header::HeaderSection

pub struct HeaderSection { /* fields omitted */ }

The complete header section of a DNS message.

Consists of a Header and a HeaderCounts.

Methods

impl HeaderSection

Creation and Conversion

fn new() -> HeaderSection

Creates a new empty header section.

fn from_message(s: &[u8]) -> &HeaderSection

Creates a reference from the bytes slice of a message.

Panics

This function panics if the size of the bytes slice is smaller than the header section.

fn from_message_mut(s: &mut [u8]) -> &mut HeaderSection

Creates a mutable reference from the bytes slice of a message.

Panics

This function panics if the size of the bytes slice is smaller than the header section.

fn as_bytes(&self) -> &[u8]

Returns a reference to the underlying bytes slice.

impl HeaderSection

Access to Header and Counts

fn header(&self) -> &Header

Returns a reference to the header.

fn header_mut(&mut self) -> &mut Header

Returns a mutable reference to the header.

fn counts(&self) -> &HeaderCounts

Returns a reference to the header counts.

fn counts_mut(&mut self) -> &mut HeaderCounts

Returns a mutable reference to the header counts.