Struct domain::bits::message::Message

pub struct Message { /* fields omitted */ }

A slice of a DNS message.

This types wraps a bytes slice with the binary content of a DNS message and allows parsing the content for further processing.

Typically, you create a message slice by passing a slice with its raw bytes to the from_bytes() function. This function only does a quick if there are enough bytes for the minimum message size. All further parsing happens lazily when you access more of the message.

Section 4 of RFC 1035 defines DNS messages as being divded into four sections named header, question, answer, authority, and additional.

The header section is of a fixed sized and can be accessed without further checks through the methods given under Header Section. Most likely, you will be interested in the first part of the header references to which are returned by the header() and header_mut() methods. The second part of the header section contains the number of entries in the following four sections and is of less interest as there are more sophisticated ways of accessing these sections. If you do care, you can get a reference through counts().

The question section contains what was asked of the DNS by a request. These questions consist of a domain name, a record type and class. With normal queries, a requests asks for all records of the given record type that are owned by the domain name within the class. There will normally be exactly one question for normal queries. With other query operations, the questions may refer to different things.

You can get access to the question section through the question() method. It returns a QuestionSection value that is an iterator over questions. Since a single question is a very common case, there is a convenience method first_question() that simple returns the first question if there is any.

The following three section all contain DNS resource records. In normal queries, they are empty in a request and may or may not contain records in a response. The answer section contains all the records that answer the given question. The authority section contains records declaring which name server provided authoritative information for the question, and the additional section can contain records that the name server thought might be useful for processing the question. For instance, if you trying to find out the mail server of a domain by asking for MX records, you likely also want the IP addresses for the server, so the name server may include these right away and free of charge.

There are functions to access all three sections directly: answer(), authority(), and additional(). However, since there are no pointers to where the later sections start, accessing them directly means iterating over the previous sections. This is why it is more efficitent to call next_section() on the returned value and process them in order. Alternatively, you can use the sections() function that gives you all four sections at once with the minimal amount of iterating necessary.

Each record in the record sections is of a specific type. Each type has its specific record data. Because there are so many types, we decided against having a giant enum. Instead, the type representing a record section, somewhat obviously named RecordSection, iterates over GenericRecords with limited options on what you can do with the data. If you are looking for a specific record type, you can get an iterator limited to records of that type through the limit_to() method. This method is generic over a record data type fit for parsing (typically meaning that it is taken from the domain::rdata::parsed module). So, if you want to iterate over the MX records in the answer section, you would do something like this:

use domain::rdata::parsed::Mx;

let msg = Message::from_bytes(bytes).unwrap();
for record in msg.answer().unwrap().limit_to::<Mx>() {
    // Do something with the record ...
}

Note that because of lazy parsing, the iterator actually returns a ParseResult<_>. One quick application of try!() fixes this:

use domain::bits::{Message, ParseResult};
use domain::rdata::parsed::Mx;

fn process_mx(msg: &Message) -> ParseResult<()> {
    for record in msg.answer().unwrap().limit_to::<Mx>() {
        let record = try!(record);
        // Do something with the record ...
    }
    Ok(())
}

Methods

impl Message

Creation and Conversion

fn from_bytes(bytes: &[u8]) -> ParseResult<&Self>

Creates a message from a bytes slice.

This fails if the slice is too short to even contain a complete header section. No further checks are done, though, so if this function returns Ok, the message may still be broken with methods returning Err(_).

unsafe fn from_bytes_unsafe(bytes: &[u8]) -> &Self

Creates a message from a bytes slice without further checks.

You need to make sure that the slice is at least the length of a full message header.

fn to_owned(&self) -> MessageBuf

Returns an owned copy of this message.

fn as_bytes(&self) -> &[u8]

Returns a reference to the underlying bytes slice.

impl Message

Header Section

fn header(&self) -> &Header

Returns a reference to the message header.

fn header_mut(&mut self) -> &mut Header

Returns a mutable reference to the message header.

The header is the only part of an already constructed message that can be safely manipulated without extra ado, so this is the only mutable method.

fn counts(&self) -> &HeaderCounts

Returns a reference to the header counts of the message.

fn no_error(&self) -> bool

Returns whether the rcode is NoError.

fn is_error(&self) -> bool

Returns whether the rcode is one of the error values.

impl Message

Sections

fn question(&self) -> QuestionSection

Returns the question section.

fn zone(&self) -> QuestionSection

Returns the zone section of an UPDATE message.

This is identical to self.question().

fn answer(&self) -> ParseResult<RecordSection>

Returns the answer section.

fn prerequisite(&self) -> ParseResult<RecordSection>

Returns the prerequisite section of an UPDATE message.

This is identical to self.answer().

fn authority(&self) -> ParseResult<RecordSection>

Returns the authority section.

fn update(&self) -> ParseResult<RecordSection>

Returns the update section of an UPDATE message.

This is identical to self.authority().

fn additional(&self) -> ParseResult<RecordSection>

Returns the additional section.

fn sections(&self)
-> ParseResult<(QuestionSection, RecordSection, RecordSection, RecordSection)>

Returns all four sections in one fell swoop.

impl Message

Helpers for Common Tasks

fn is_answer<M: AsRef<Message>>(&self, query: M) -> bool

Returns whether this is the answer to some other message.

The method checks whether the ID fields of the headers are the same, whether the QR flag is set in this message, and whether the questions are the same.

fn first_question(&self) -> Option<Question<ParsedDName>>

Returns the first question, if there is any.

The method will return None both of there are no questions or if parsing fails.

fn qtype(&self) -> Option<Rtype>

Returns the query type of the first question, if any.

fn contains_answer<'a, D: ParsedRecordData<'a>>(&'a self) -> bool

Returns whether the message contains answers of a given type.

fn canonical_name(&self) -> Option<ParsedDName>

Resolves the canonical name of the answer.

Returns None if either the message doesn’t have a question or there was a parse error. Otherwise starts with the question’s name, follows any CNAME trail and returns the name answers should be for.