Struct domain::bits::message::MessageBuf

pub struct MessageBuf { /* fields omitted */ }

An owned DNS message.

This type owns the underlying bytes of the message and derefs into a Message for all processing. For more information on DNS messages and how they can be accessed, please refer to the documentation of the Message type.

This is, however, not the type for building messages. Use [MessageBuilder] instead.

Methods

impl MessageBuf

Creation and Conversion

fn from_vec(vec: Vec<u8>) -> ParseResult<Self>

Creates a new owned message using the given vector.

If the content of the vector is too short to even contain a full header, the function fails.

fn from_bytes(slice: &[u8]) -> ParseResult<Self>

Creates a new owned message cloning the data from the bytes slice.

If the slice is too short to even contain a full header section, the function fails.

fn as_slice(&self) -> &Message

Returns a reference to the message slice.