Enum domain::bits::name::FromStrError

pub enum FromStrError {
    UnexpectedEnd,
    EmptyLabel,
    LongLabel,
    IllegalEscape,
    IllegalCharacter,
    IllegalBinary,
    RelativeName,
    LongName,
}

Variants
UnexpectedEnd

The string ended when there should have been more characters.

This most likely happens inside escape sequences and quoting.

EmptyLabel

An empty label was encountered.

LongLabel

A domain name label has more than 63 octets.

IllegalEscape

An illegal escape sequence was encountered.

Escape sequences are a backslash character followed by either a three decimal digit sequence encoding a byte value or a single other printable ASCII character.

IllegalCharacter

An illegal character was encountered.

Only printable ASCII characters are allowed.

IllegalBinary

An illegal binary label sequence was encountered.

RelativeName

A relative name was encountered.

LongName

The name has more than 255 characters.