Enum domain::bits::name::Labelette

pub enum Labelette<'a> {
    Normal(&'a [u8]),
    Bit(bool),
}

A labelette is an atomic label.

Various operations on domain names such as comparisons or ordering are done label by label. However, binary labels are actually a sequence of labels themselves. Worse, a binary label within a name can actually be broken up into several binary labels with the same overall number of bits while still retaining the same domain name. All these operations need to consider the individual bits labels of a binary label.

In order to disperse the confusion of having labels inside labels, we invented the term labelette exclusively for this crate to mean either a normal label or one single bit-label of a binary label.

You get labelettes by iterating over Labels. For a normal label, this iterator will return exactly once with a Normal variant. For binary labels, the iterator will return once for each bit label with the Bit variant.

Variants
Normal(&'a [u8])

A labelette for a normal label.

Bit(bool)

A labelette for a single bit-label of a binary label.

Methods

impl<'a> Labelette<'a>

fn is_root(&self) -> bool

Returns whether the labelette is the root label.