Struct domain::bits::name::DNameBuf

pub struct DNameBuf { /* fields omitted */ }

An owned complete domain name.

A value of this type contains a vector with a correctly encoded, uncompressed domain name. It derefs to DNameSlice in order to make all its methods available for working with domain names.

In addition, it provides a number of methods to add labels or entire names to its end.

DNameBuf values can be created from string via the std::str::FromStr trait. Such strings must be in the usual zonefile encoding.

Methods

impl DNameBuf

Creation and Conversion

fn new() -> Self

Creates a new empty domain name.

fn with_capacity(capacity: usize) -> DNameBuf

Creates a new empty name with the given amount of space reserved.

fn from_vec(vec: Vec<u8>) -> Option<Self>

Creates an owned domain name using an existing bytes vector.

If the content of the bytes vector does not constitute a correctly encoded uncompressed domain name, the function will fail.

fn from_iter<'a, T>(iter: T) -> Result<Self, PushError> where T: IntoIterator<Item=&'a Label>

Creates an owned domain name from the labels of the iterator.

fn scan<S: Scanner>(scanner: &mut S,
origin: Option<&DNameSlice>)
-> ScanResult<Self>

Creates an owned domain name by reading it from a scanner.

fn root() -> DNameBuf

Returns a new owned domain name consisting only of the root label.

fn as_slice(&self) -> &DNameSlice

Returns a reference to a slice of the domain name.

fn into_vec(self) -> Vec<u8>

Extracts the underlying vector from the name.

impl DNameBuf

Manipulations

fn push(&mut self, label: &Label) -> Result<(), PushError>

Extends a relative name with a label.

If the name is absolute, nothing happens.

fn push_normal(&mut self, content: &[u8]) -> Result<(), PushError>

Pushes a normal label to the end of a relative name.

If the name is absolute, nothing happens. If the resulting name would exceed the maximum allowd length of 255 octets, returns an error.

Panics

The method panics if content is longer that 63 bytes.

fn push_binary(&mut self, count: usize, bits: &[u8]) -> Result<(), PushError>

Pushes a binary label to the end of a relative name.

The binary label will be count bits long and contain the bits from bits. If bits is too short, the label will be filled up with zero bits. If bits is too long, it will be trimmed to the right length.

If the name is absolute, nothing happens. If the resulting name would exceed the maximum allowd length of 255 octets, returns an error.

Panics

The method panics if count is larger than 256.

fn push_empty_binary(&mut self,
count: usize)
-> Result<Option<&mut [u8]>, PushError>

Pushes an empty binary label of the given length to the domain name.

Upon success the function returns a mutable reference to the bytes slice of the bits of the binary label.

If the name is already absolute, returns Ok(None). If the resulting name would exceed the maximum allowd length of 255 octets, returns an error.

Panics

The method panics if count is larger than 256.

fn append<N: DName>(&mut self, name: &N) -> Result<(), PushError>

Extends a relative name with a domain name.

If the name is already absolute, nothing will be appended and the name remains unchanged. If by appending the name would exceed the maximum allowed length of 255 octets, an error will be returned and the name remains unchanged, too.

fn append_root(&mut self) -> Result<(), PushError>

Makes a domain name absolute if it isn’t yet by appending root.

This may fail if the name is already 255 octets long.

fn append_iter<'a, T>(&mut self, iter: T) -> Result<(), PushError> where T: IntoIterator<Item=&'a Label>

Appends the content of an iterator to the end of the name.

fn remove_first(&mut self)

Removes the first label.

After this operation, the content of self is self.parent().

Panics

Panics if the name is empty.