Struct domain::bits::name::LabelBuf

pub struct LabelBuf { /* fields omitted */ }

An owned domain name label.

This type is the owned companion of and derefs to [Label].

Methods

impl LabelBuf

fn from_slice(label: &Label) -> Self

Creates an owned label from a label slice.

LabelBuf also implements From<&Label>, so the canonical form of this function is actually label.into().

fn root() -> Self

Returns an owned root label.

fn as_slice(&self) -> &Label

Returns a reference to a slice of the label.