Trait domain::bits::rdata::RecordData

pub trait RecordData: Sized {
    fn rtype(&self) -> Rtype;
    fn compose<C: AsMut<Composer>>(&self, target: C) -> ComposeResult<()>;
}

A trait for types representing record data.

Required Methods

fn rtype(&self) -> Rtype

Returns the record type for this record data instance.

This is a method rather than an associated function to allow one type to be used for several real record types.

fn compose<C: AsMut<Composer>>(&self, target: C) -> ComposeResult<()>

Appends the record data to the end of a composer.