Struct domain::bits::record::Record

pub struct Record<N: DName, D: RecordData> { /* fields omitted */ }

A DNS resource record.

All information available through the DNS is stored in resource records. They have a three part key of a domain name, resource record type, and class. Data is arranged in a tree which is navigated using the domain name. Each node in the tree carries a label, starting with the root label as the top-most node. The tree is traversed by stepping through the name from right to left, finding a child node carring the label of each step.

The record type describes the kind of data the record holds, such as IP addresses. The class, finally, describes which sort of network the information is for since DNS was originally intended to be used for networks other than the Internet as well. In practice, the only relevant class is IN, the Internet. Note that each class has its own tree of nodes.

The payload of a resource record is its data. Its purpose, meaning, and format is determined by the record type. For each unique three-part key there can be multiple resource records. All these records for the same key are called resource record sets, most often shortened to ‘RRset.’

There is one more piece of data: the TTL or time to live. This value says how long a record remains valid before it should be refreshed from its original source, given in seconds. The TTL is used to add caching facilities to the DNS.

Values of the Record type represent one single resource record. Since there are currently more than eighty record types—see Rtype for a complete list—, the type is generic over a trait for record data. This trait holds both the record type value and the record data as they are inseparably entwined.

Since records contain domain names, the Record type is additionally generic over a domain name type.

There is three ways to create a record value. First, you can make one yourself using the [new()] function. In will neatly take care of all the generics through type inference. Secondly, you can parse a record from an existing message. Message and its friends provide a way to do that; see there for all the details. Finally, you can scan a record from master data (aka zonefiles). See the domain::master module for that.

Records can be place into DNS messages through the value chain starting with MessageBuilder. In order to make adding records easier, Record implements the From trait for two kinds of tuples: A four-tuple of name, class, time-to-live value, and record data and a triple leaving out the class and assuming it to Class::In.

Methods

impl<N: DName, D: RecordData> Record<N, D>

Creation and Element Access

fn new(name: N, class: Class, ttl: u32, data: D) -> Self

Creates a new record from its parts.

fn name(&self) -> &N

Returns a reference to the domain name.

The domain name, sometimes called the owner of the record, specifies the node in the DNS tree this record belongs to.

fn rtype(&self) -> Rtype

Returns the record type.

fn class(&self) -> Class

Returns the record class.

fn set_class(&mut self, class: Class)

Sets the record’s class.

fn ttl(&self) -> u32

Returns the record’s time-to-live.

fn set_ttl(&mut self, ttl: u32)

Sets the record’s time-to-live.

fn data(&self) -> &D

Return a reference to the record data.

fn data_mut(&mut self) -> &mut D

Returns a mutable reference to the record data.

fn into_data(self) -> D

Trades the record for its record data.

impl<'a, D: ParsedRecordData<'a>> Record<ParsedDName<'a>, D>

Parsing

fn parse(parser: &mut Parser<'a>) -> ParseResult<Option<Self>>

Parses a record from a parser.

This function is only available for records that use a parsed domain name and a record data type that, too, can be parsed.

impl<'a> Record<ParsedDName<'a>, GenericRecordData<'a>>

fn parse_generic(parser: &mut Parser<'a>) -> ParseResult<Self>

Parses a record with generic data from a parser.

impl<N: DName, D: RecordData> Record<N, D>

Composing

fn compose<C: AsMut<Composer>>(&self, composer: C) -> ComposeResult<()>

Appends the record’s wire-format representation to a composer.