Struct domain::bits::DNameSlice

pub struct DNameSlice { /* fields omitted */ }

A slice of a domain name.

The slice is guaranteed to contain a correctly encoded domain name. The name may be relative or absolute but cannot support name compression.

Operations are available for iterating over the labels of the name and breaking the name up into parts along the lines of label boundaries.

This is an unsized type. You will have to use it with some kind of pointer, such as a reference or box. The sibling type owning the name outright is DNameBuf. The two can be used together through the AsRef<DNameSlice> trait or via a Cow<DNameSlice>.

Methods

impl DNameSlice

Creation and Convertion

unsafe fn from_bytes_unsafe(bytes: &[u8]) -> &DNameSlice

Creates a domain name slice from a bytes slice without checking.

This is only safe if the input slice follows the encoding rules for a domain name and does not contain compressed labels.

fn from_bytes(bytes: &[u8]) -> Option<&Self>

Creates a domain name from a bytes slice if it contains a valid name.

The bytes slice must be exactly one correctly encoded, uncompressed domain name. If it isn’t or if it contains trailing data, the function returns None.

As this function traverses the slice to check that it is correctly encoded, it may take a bit of time.

fn root() -> &'static DNameSlice

Creates a domain name slice for the root domain name.

fn empty() -> &'static DNameSlice

Creates an empty domain name slice.

fn as_bytes(&self) -> &[u8]

Returns a reference to the underlying bytes slice.

fn to_owned(&self) -> DNameBuf

Converts the domain name slice to an owned domain name.

impl DNameSlice

Properties

fn is_absolute(&self) -> bool

Checks whether the domain name is absolute.

A domain name is absolute if it ends with an empty normal label (the root label).

fn is_relative(&self) -> bool

Checks whether the domain name is relative, ie., not absolute.

impl DNameSlice

Iterating over Labels

fn labels(&self) -> NameLabels

Produces an iterator over the labels in the name.

fn labelettes(&self) -> NameLabelettes

Produces an iterator over the labelettes in the name.

fn len(&self) -> usize

Returns the number of labels in self.

fn is_empty(&self) -> bool

Checks whether the domain name is empty.

fn first(&self) -> Option<&Label>

Returns the first label or None if the name is empty.

fn last(&self) -> Option<&Label>

Returns the last label or None if the name is empty.

fn ndots(&self) -> Option<usize>

Returns the number of dots if this is a relative name.

Returns None if this is an absolute name.

impl DNameSlice

Working with Parts

fn split_first(&self) -> Option<(&Label, &Self)>

Returns the first label and the remaining domain name.

Returns None only if the name is empty (which is different from a name containing only the root label).

fn parent(&self) -> Option<&DNameSlice>

Returns the domain name without its leftmost label.

Returns None for an empty domain name and a domain name consisting of only the root label. Returns an empty domain name for a single label relative domain name.

fn strip_suffix<'a, N: DName>(&'a self,
base: &'a N)
-> Result<Cow<'a, Self>, StripSuffixError>

Returns a domain name slice relative to base.

This fails if base isn’t a suffix of self.

fn starts_with<N: DName>(&self, base: &N) -> bool

Determines whether base is a prefix of self.

fn ends_with<N: DName>(&self, base: &N) -> bool

Determines whether base is a suffix of self.

fn join<N: DName>(&self, base: &N) -> Result<DNameBuf, PushError>

Creates an owned domain name made absolute if necessary.

If self is already an absolute domain name, nothing happens.