Enum domain::iana::opcode::Opcode

pub enum Opcode {
    Query,
    IQuery,
    Status,
    Notify,
    Update,
    Int(u8),
}

DNS OpCodes.

The opcode specifies the kind of query to be performed.

The opcode is initially defined in RFC 1035. All currently assigned values can be found at http://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-5

Variants
Query

A standard query.

This query requests all records matching the name, class, and record type given in the query’s question section.

This value is defined in RFC 1035.

IQuery

An inverse query (IQUERY) (obsolete).

The idea behind inverse queries was to provide a single answer and ask the DNS for all the questions that would lead to this answer. This kind of query has always been optional, was never widely supported, and has therefore been declared obsolete.

This value was defined in RFC 1035 and obsoleted by RFC 3425.

Status

A server status request.

This value is defined in RFC 1035. The status request itself was defined as experimental and ‘to be defined’ in RFC 1034 and seems to never have been mentioned ever again.

Notify

A NOTIFY query.

NOTIFY queries allow master servers to inform slave servers when a zone has changed.

This value and the NOTIFY query are defined in RFC 1996.

Update

An UPDATE query.

The UPDATE query can be used to alter zone content managed by a master server.

This value and the UPDATE query are defined in RFC 2136.

Int(u8)

A raw integer opcode value.

When converting to an u8, only the lower four bits are used.

Methods

impl Opcode

fn from_int(value: u8) -> Opcode

Creates an Opcode value from an integer value.

Only considers the lower four bits of value.

fn to_int(self) -> u8

Returns the integer value for this opcode.