Enum domain::iana::rcode::OptRcode

pub enum OptRcode {
    NoError,
    FormErr,
    ServFail,
    NXDomain,
    NotImp,
    Refused,
    YXDomain,
    YXRRSet,
    NXRRSet,
    NotAuth,
    NotZone,
    BadVers,
    BadCookie,
    Int(u16),
}

Extended DNS Response Codes for OPT records.

Originally, the response code of embedded in the header of each DNS message was four bits long. This code, defined in [RFC 1035], is represented by the Rcode type. The extension mechanism for DNS initially defined in RFC 2671 and updated by RFC 6891 added eight more bits to be stored in the OPT pseudo-resource record. This type represents the complete 12 bit extended response code.

There is a third, 16 bit wide response code for transaction authentication (TSIG) defined in RFC 2845 and represented by the TsigRcode type. The code mostly shares the same name space except for an unfortunate collision in between the BADVERS and BADSIG values. Because of this, we decided to have separate types.

The values for all three response code types are defined in the IANA DNS RCODEs registry.

Variants
NoError

No error condition.

(Otherwise known as success.)

Defined in RFC 1035.

FormErr

Format error.

The name server was unable to interpret the query.

Defined in RFC 1035.

ServFail

Server failure.

The name server was unable to process this query due to a problem with the name server.

Defined in RFC 1035.

NXDomain

Name error.

The domain name given in the query does not exist at the name server.

Defined in RFC 1035.

NotImp

Not implemented.

The name server does not support the requested kind of query.

Defined in RFC 1035.

Refused

Query refused.

The name server refused to perform the operation requested by the query for policy reasons.

Defined in RFC 1035.

YXDomain

Name exists when it should not.

Returned for an UPDATE query when a domain requested to not exist does in fact exist.

Returned when resolving a DNAME redirection when the resulting name exceeds the length of 255 octets.

Defined in RFC 2136 for the UPDATE query and RFC 6672 for DNAME redirection.

YXRRSet

RR set exists when it should not.

Returned for an UPDATE query when an RRset requested to not exist does in fact exist.

Defined in RFC 2136.

NXRRSet

RR set that should exist does not.

Returned for an UPDATE query when an RRset requested to exist does not.

Defined in RFC 2136.

NotAuth

Server not authoritative for zone or client not authorized.

Returned for an UPDATE query when the server is not an authoritative name server for the requested domain.

Returned for queries using TSIG when authorisation failed.

Defined in RFC 2136 for UPDATE and RFC 2845 for TSIG.

NotZone

Name not contained in zone.

A name used in the prerequisite or update section is not within the zone given in the zone section.

Defined in RFC 2136.

BadVers

Bad OPT version.

A name server does not implement the EDNS version requested in the OPT record.

Defined in RFC 6891.

BadCookie

Bad or missing server cookie.

The request contained a COOKIE option either without a server cookie or with a server cookie that did not validate.

Defined in RFC 7873.

Int(u16)

A raw, integer rcode value.

When converting to a 12 bit code, the upper four bits are simply ignored.

Methods

impl OptRcode

fn from_int(value: u16) -> OptRcode

Creates an rcode from an integer.

Only the lower twelve bits of value are considered.

fn to_int(self) -> u16

Returns the integer value for this rcode.

fn from_parts(rcode: Rcode, ext: u8) -> OptRcode

Creates an extended rcode value from its parts.

fn to_parts(self) -> (Rcode, u8)

Returns the two parts of an extended rcode value.

fn rcode(&self) -> Rcode

Returns the rcode part of the extended rcode.

fn ext(&self) -> u8

Returns the extended octet of the extended rcode.