Enum domain::iana::rcode::TsigRcode

pub enum TsigRcode {
    NoError,
    FormErr,
    ServFail,
    NXDomain,
    NotImp,
    Refused,
    YXDomain,
    YXRRSet,
    NXRRSet,
    NotAuth,
    NotZone,
    BadSig,
    BadKey,
    BadTime,
    BadMode,
    BadName,
    BadAlg,
    BadTrunc,
    BadCookie,
    Int(u16),
}

Response codes for transaction authentication (TSIG).

TSIG and TKEY resource records contain a 16 bit wide error field whose values are an extension of the standard DNS [Rcode]. While it was intended to also share the same space with the extended response codes used by EDNS (see OptRcode), both used the value 16. To allow distinguish between the two uses of this value, we have two separate types.

The values for all three response code types are defined in the IANA DNS RCODEs registry.

Variants
NoError

No error condition.

(Otherwise known as success.)

Defined in RFC 1035.

FormErr

Format error.

The name server was unable to interpret the query.

Defined in RFC 1035.

ServFail

Server failure.

The name server was unable to process this query due to a problem with the name server.

Defined in RFC 1035.

NXDomain

Name error.

The domain name given in the query does not exist at the name server.

Defined in RFC 1035.

NotImp

Not implemented.

The name server does not support the requested kind of query.

Defined in RFC 1035.

Refused

Query refused.

The name server refused to perform the operation requested by the query for policy reasons.

Defined in RFC 1035.

YXDomain

Name exists when it should not.

Returned for an UPDATE query when a domain requested to not exist does in fact exist.

Returned when resolving a DNAME redirection when the resulting name exceeds the length of 255 octets.

Defined in RFC 2136 for the UPDATE query and RFC 6672 for DNAME redirection.

YXRRSet

RR set exists when it should not.

Returned for an UPDATE query when an RRset requested to not exist does in fact exist.

Defined in RFC 2136.

NXRRSet

RR set that should exist does not.

Returned for an UPDATE query when an RRset requested to exist does not.

Defined in RFC 2136.

NotAuth

Server not authoritative for zone or client not authorized.

Returned for an UPDATE query when the server is not an authoritative name server for the requested domain.

Returned for queries using TSIG when authorisation failed.

Defined in RFC 2136 for UPDATE and RFC 2845 for TSIG.

NotZone

Name not contained in zone.

A name used in the prerequisite or update section is not within the zone given in the zone section.

Defined in RFC 2136.

BadSig

TSIG signature failure.

The TSIG signature fails to verify.

Defined in RFC 2845.

BadKey

Key not recognized.

The server did not recognize the key used for generating the signature.

Defined in RFC 2845.

BadTime

Signature out of time window.

The server time was outside the time interval specified by the request.

Defined in RFC 2845.

BadMode

Bad TKEY mode.

The mode field in a TKEY resource record contained a mode not supported by the server.

Defined in RFC 2930.

BadName

Duplicate key name.

In TKEY records, when establishing a new key, the name used already exists at the server or when deleting a key, a key of this name does not exist.

Defined in RFC 2930.

BadAlg

Algorithm not supported.

The value is defined in RFC 2930 but never actually explained. Presumably, it will be returned when the algorithm field of a TKEY record contains a value not supported by the server.

BadTrunc

Bad truncation.

A TSIG record was received with a MAC too short for the local policy in force.

Defined in RFC 4635.

BadCookie

Bad or missing server cookie.

The request contained a COOKIE option either without a server cookie or with a server cookie that did not validate.

Defined in RFC 7873.

Int(u16)

A raw, integer rcode value.

When converting to a 12 bit code, the upper four bits are simply ignored.

Methods

impl TsigRcode

fn from_int(value: u16) -> TsigRcode

Creates an rcode from an integer.

fn to_int(self) -> u16

Returns the integer value for this rcode.