Enum domain::iana::rtype::Rtype

pub enum Rtype {
    A,
    Ns,
    Md,
    Mf,
    Cname,
    Soa,
    Mb,
    Mg,
    Mr,
    Null,
    Wks,
    Ptr,
    Hinfo,
    Minfo,
    Mx,
    Txt,
    Rp,
    Afsdb,
    X25,
    Isdn,
    Rt,
    Nsap,
    Nsapptr,
    Sig,
    Key,
    Px,
    Gpos,
    Aaaa,
    Loc,
    Nxt,
    Eid,
    Nimloc,
    Srv,
    Atma,
    Naptr,
    Kx,
    Cert,
    A6,
    Dname,
    Sink,
    Opt,
    Apl,
    Ds,
    Sshfp,
    Ipseckey,
    Rrsig,
    Nsec,
    Dnskey,
    Dhcid,
    Nsec3,
    Nsec3param,
    Tlsa,
    Smimea,
    Hip,
    Ninfo,
    Rkey,
    Talink,
    Cds,
    Cdnskey,
    Openpgpkey,
    Csync,
    Spf,
    Uinfo,
    Uid,
    Gid,
    Unspec,
    Nid,
    L32,
    L64,
    Lp,
    Eui48,
    Eui64,
    Tkey,
    Tsig,
    Ixfr,
    Axfr,
    Mailb,
    Maila,
    Any,
    Uri,
    Caa,
    Avc,
    Ta,
    Dlv,
    Int(u16),
}

Resource Record Types.

Each resource records has a 16 bit type value indicating what kind of information is represented by the record. Normal query includes the type of record information is requested for. A few aditional types, called query types, are defined as well and can only be used in questions. This type represents both these types.

Record types are defined in RFC 1035. The registry of currently assigned values can be found at http://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-4

In order to avoid confusion over capitalization, the mnemonics are treated as single acronyms and therefore all variant names are spelled with an initial capital letter in accordance with Rust naming guidelines.

Variants
A

A host address.

Ns

An authoritative name server.

Md

A mail destination.

(Obsolete – use MX)

Mf

A mail forwarder.

(Obsolete – use MX)

Cname

The canonical name for an alias

Soa

Marks the start of a zone of authority.

Mb

A mailbox domain name.

(Experimental.)

Mg

A mail group member

(Experimental.)

Mr

A mail rename domain name.

(Experimental.)

Null

A null resource record.

(Experimental.)

Wks

A well known service description.

Ptr

A domain name pointer.

Hinfo

Host information.

Minfo

Mailbox or mail list information.

Mx

Mail exchange.

Txt

Text strings.

Rp

For Responsible Person.

See RFC 1183

Afsdb

For AFS Data Base location.

See RFC 1183 and RFC 5864.

X25

For X.25 PSDN address.

See RFC 1183.

Isdn

For ISDN address.

See RFC 1183.

Rt

For Route Through.

See RFC 1183

Nsap

For SNAP address, NSAP style A record.

See RFC 1706.

Nsapptr

For domain name pointer, NSAP style.

See RFC 1348, RFC 1637, RFC 1706.

Sig

For security signature.

Key

For security key.

Px

X.400 mail mapping information.

See RFC 2163.

Gpos

Geographical position.

See RFC 1712

Aaaa

IPv6 address.

See RFC 3596.

Loc

Location information.

See RFC 1876.

Nxt

Next domain.

(Obsolete.)

See RFC 3755 and RFC 2535.

Eid

Endpoint identifier.

Nimloc

Nimrod locator.

Srv

Server selection.

See RFC 2782.

Atma

ATM address.

Naptr

Naming authority pointer.

See RFC 2915, RFC 2168, and RFC 3403.

Kx

Key exchanger.

See RFC 2230.

Cert

CERT

See RFC 4398.

A6

A6.

(Obsolete – use AAAA.)

See RFC 3226, RFC 2874, and RFC 6563.

Dname

DNAME.

See RFC 6672.

Sink

SINK.

Opt

OPT.

See RFC 6891 and RFC 3225.

Apl

APL.

See RFC 3123.

Ds

Delegation signer.

See RFC 4034 and RFC 3658.

Sshfp

SSH key fingerprint.

See RFC 4255.

Ipseckey

IPSECKEY

See RFC 4255.

Rrsig

RRSIG.

See RFC 4034 and RFC 3755.

Nsec

NSEC.

See RFC 4034 and RFC 3755.

Dnskey

DNSKEY.

See RFC 4034 and RFC 3755.

Dhcid

DHCID.

See RFC 4701.

Nsec3

NSEC3

See RFC 5155.

Nsec3param

NSEC3PARAM.

See RFC 5155.

Tlsa

TLSA.

See RFC 6698.

Smimea

S/MIME cert association.

See draft-ietf-dane-smime.

Hip

Host Identity Protocol.

See RFC 5205.

Ninfo

NINFO.

Rkey

RKEY.

Talink

Trust Anchor Link

Cds

Child DS.

See RFC 7344.

Cdnskey

DNSKEY(s) the child wants reflected in DS.

See RFC 7344.

Openpgpkey

OpenPGP key.

See draft-ietf-dane-openpgpkey.

Csync

Child-to-parent synchronization.

See RFC 7477.

Spf

SPF.

RFC 7208.

Uinfo

UINFO.

IANA-Reserved.

Uid

UID.

IANA-Reserved.

Gid

GID.

IANA-Reserved.

Unspec

UNSPEC.

IANA-Reserved.

Nid

NID.

See RFC 6742.

L32

L32.

See RFC 6742.

L64

L64.

See RFC 6742.

Lp

LP.

See RFC 6742.

Eui48

An EUI-48 address.

See RFC 7043.

Eui64

An EUI-64 address.

See RFC 7043.

Tkey

Transaction key.

See RFC 2930.

Tsig

Transaction signature.

See RFC 2845.

Ixfr

Incremental transfer.

See RFC 1995.

Axfr

Transfer of entire zone.

See RFC 1035 and RFC 5936.

Mailb

Mailbox-related RRs (MB, MG, or MR).

Maila

Mail agent RRS.

(Obsolete – see MX.)

Any

A request for all records the server/cache has available.

See RFC 1035 and RFC 6895.

Uri

URI.

See RFC 7553.

Caa

Certification Authority Restriction.

See RFC 6844.

Avc

Application visibility and control.

Ta

DNSSEC trust authorities.

Dlv

DNSSEC lookaside validation.

See RFC 4431

Int(u16)

A raw value given through its integer.

Methods

impl Rtype

fn from_int(value: u16) -> Self

Returns a value from its raw integer value.

fn to_int(self) -> u16

Returns the raw integer value for a value.

fn from_mnemonic(m: &[u8]) -> Option<Self>

Returns a value from a well-defined mnemonic.

fn to_mnemonic(self) -> Option<&'static [u8]>

Returns the mnemonic for this value if there is one.

This will also return a mnemonic if a well-defined variant is hidden in a Int variant.

impl Rtype

fn from_bytes(bytes: &[u8]) -> Option<Self>

impl Rtype

fn parse(parser: &mut Parser) -> ParseResult<Self>

fn compose<C: AsMut<Composer>>(&self, composer: C) -> ComposeResult<()>

fn scan<S: Scanner>(scanner: &mut S) -> ScanResult<Self>