Enum domain::iana::secalg::SecAlg

pub enum SecAlg {
    RsaMd5,
    Dh,
    Dsa,
    RsaSha1,
    DsaNsec3Sha1,
    RsaSha1Nsec3Sha1,
    RsaSha256,
    RsaSha512,
    EccGost,
    EcdsaP256Sha256,
    EcdsaP384Sha384,
    Indirect,
    PrivateDns,
    PrivateOid,
    Int(u8),
}

See http://www.iana.org/assignments/dns-sec-alg-numbers/dns-sec-alg-numbers.xhtml#dns-sec-alg-numbers-1

Variants
RsaMd5

RSA/MD5

This algorithm was described in RFC 2537 and since has been deprecated due to weaknesses of the MD5 hash algorithm by RFC 3110 which suggests to use RSA/SHA1 instead.

This algorithm may not be used for zone signing but may be used for transaction security.

Dh

Diffie-Hellman

This algorithm is described in RFC 2539 for storing Diffie-Hellman (DH) keys in DNS resource records. It can not be used for zone signing but only for transaction security.

Dsa

DSA/SHA1

This algorithm is described in RFC 2536. It may be used both for zone signing and transaction security.

RsaSha1

RSA/SHA-1

This algorithm is described in RFC 3110. It may be used both for zone signing and transaction security. It is mandatory for DNSSEC implementations.

DsaNsec3Sha1

DSA-NSEC3-SHA1

This value is an alias for Dsa for use within NSEC3 records.

RsaSha1Nsec3Sha1

RSASHA1-NSEC3-SHA1

This value is an alias for RsaSha1 for use within NSEC3 records.

RsaSha256

RSA/SHA-256

This algorithm is described in RFC 5702. It may be used for zone signing only.

RsaSha512

RSA/SHA-512

This algorithm is described in RFC 5702. It may be used for zone signing only.

EccGost

GOST R 34.10-2001

This algorithm is described in RFC 5933. It may be used for zone signing only.

EcdsaP256Sha256

ECDSA Curve P-256 with SHA-256

This algorithm is described in RFC 6605. It may be used for zone signing only.

EcdsaP384Sha384

ECDSA Curve P-384 with SHA-384

This algorithm is described in RFC 6605. It may be used for zone signing only.

Indirect

Reserved for Indirect Keys

This value is reserved by RFC 4034.

PrivateDns

A private algorithm identified by a domain name.

This value is defined in RFC 4034.

PrivateOid

A private algorithm identified by a ISO OID.

This value is defined in RFC 4034.

Int(u8)

A raw algorithm value given through its integer value.

Methods

impl SecAlg

fn from_int(value: u8) -> SecAlg

Returns the algorithm value for the given integer value.

fn to_int(self) -> u8

Returns the integer value for this algorithm value.