Struct domain::master::bufscanner::BufScanner

pub struct BufScanner<R: Read> { /* fields omitted */ }

A scanner using a buffer atop a reader.

The strategy implemented for the buffer is to read byte by byte from the underlying reader and flush it every time a successfully read token ends at the buffer’s end (which should happen quite a lot). This may not be the smartest way of doing this, but it is simple and memory-efficient.

Methods

impl<R: Read> BufScanner<R>

Creation

fn new(reader: R) -> Self

Creates a new scanner using the given reader.

fn with_pos(reader: R, pos: Pos) -> Self

Creates a new scanner using the given reader and position.

The scanner will assume that the current position of reader corresponds to the human-friendly position pos.

impl BufScanner<File>

fn open<P: AsRef<Path>>(path: P) -> Result<Self>

Attempts to opens a file and create a scanner for it.

impl<T: AsRef<[u8]>> BufScanner<Cursor<T>>

fn create(t: T) -> Self

Wraps t into a cursor and creates a scanner with it.