Module domain::rdata

Resource data implementations.

This module will eventually contain implementations for the record data for all defined resource record types.

The types are named identically to the RRType variant they implement. They are grouped into submodules for the RFCs they are defined in. All types are also re-exported at the top level here. Ie., for the AAAA record type, you can simple use domain::rdata::Aaaa instead of use domain::rdata::rfc3596::Aaaa which nobody could possibly remember. There are, however, some helper data types defined here and there which are not re-exported to keep things somewhat tidy.

See the RRType enum for the complete set of record types and, consequently, those types that are still missing.

Reexports

pub use self::rfc1035::A;

pub use self::rfc1035::Cname;

pub use self::rfc1035::Hinfo;

pub use self::rfc1035::Mb;

pub use self::rfc1035::Md;

pub use self::rfc1035::Mf;

pub use self::rfc1035::Mg;

pub use self::rfc1035::Minfo;

pub use self::rfc1035::Mr;

pub use self::rfc1035::Mx;

pub use self::rfc1035::Ns;

pub use self::rfc1035::Ptr;

pub use self::rfc1035::Soa;

pub use self::rfc1035::Txt;

pub use self::rfc1035::Wks;

pub use self::rfc3596::Aaaa;

pub use self::rfc1035::Null;

Modules

owned

parsed

rfc1035
Record data from RFC 1035.

rfc3596
Record data from RFC 3596.

Enums

MasterRecordData

Functions

fmt_rdata