Module domain::rdata::rfc1035

Record data from RFC 1035.

This RFC defines the initial set of record types.

Modules

owned

parsed

Structs

A
A record data.

Cname

Hinfo
Hinfo record data.

Mb

Md

Mf

Mg

Minfo
Minfo record data.

Mr

Mx
Mx record data.

Ns

Null
Null record data.

Ptr

Soa
Soa record data.

Txt
Txt record data.

TxtIter
An iterator over the character strings of a Txt record.

Wks
Wks record data.

WksBitmap

WksBitmapBuf

WksIter
An iterator over the services active in a Wks record.