Struct domain::rdata::rfc1035::Wks

pub struct Wks<B: AsRef<WksBitmap>> { /* fields omitted */ }

Wks record data.

Wks records describe the well-known services supported by a particular protocol on a particular internet address.

The Wks record type is defined in RFC 1035, section 3.4.2.

Methods

impl<B: AsRef<WksBitmap>> Wks<B>

fn new(address: Ipv4Addr, protocol: u8, bitmap: B) -> Self

Creates a new record data from components.

fn address(&self) -> Ipv4Addr

The IPv4 address of the host this record refers to.

fn protocol(&self) -> u8

The protocol number of the protocol this record refers to.

This will typically be 6 for TCP or 17 for UDP.

fn bitmap(&self) -> &WksBitmap

A bitmap indicating the ports where service is being provided.

fn serves(&self, port: u16) -> bool

Returns whether a certain service is being provided.

fn iter(&self) -> WksIter

Returns an iterator over the served ports.

impl Wks<WksBitmapBuf>

fn scan<S: Scanner>(scanner: &mut S,
origin: Option<&DNameSlice>)
-> ScanResult<Self>