Module domain::resolv::conf

Resolver configuration

There are two parts to this module: Query options that allow you to modify the behaviour of the resolver on a query by query basis and the global resolver configuration (normally read from the system’s /etc/resolv.conf) that contains things like the name servers to query and a set of default options.

Both parts are modeled along the lines of glibc’s resolver.

The resolver configuration is going to be refactored soon to be more in line with the usage of the resolver.

Structs

ResolvConf
Resolver configuration.

ResolvOptions
Options for the resolver configuration.

ServerConf
Configuration for one upstream DNS server.

Enums

Error
The error that can happen when parsing resolv.conf.

ServerMode
The mode of operation of a server over a given transport.

Type Definitions

Result
The result from parsing resolv.conf.