Struct domain::resolv::lookup::addr::FoundAddrs

pub struct FoundAddrs(_);

The success type of the lookup_addr() function.

The only purpose of this type is to return an iterator over host names via its iter() method.

Methods

impl FoundAddrs

fn iter(&self) -> FoundAddrsIter

Returns an iterator over the host names.